Tests for `measurement.value_as_number`
Related to #1

- New seed file with lower and upper limit columns for each measurement_concept_id currently in the RXN OMOP database
- These values default to the minimum and maximum values in the database - but these are likely to be the wrong limits
- When these values have been modified, this is noted by the initials of the person modifying them in the `verified_by` column

A test macro compares the values in the `value_as_number` column against the limits in the seed model by joining on `measurement_concept_id` and returns rows where values fall outside of this range.

Needs more work to update all the limits in the CSV file which needs to be kept updated when new concepts appear in the MEASUREMENT table.